Revert "Use WikiPage::prepareContentForEdit in SpamBlacklistHooks::filterMergedContent"
For Wikidata, this causes users to be unable to create items. (bug 59797)

In case of Special:NewItem, $context->getWikiPage() does
$context->getTitle() to get the special page and then WikiPage::factory
with the special page $title.  special page does not make a valid
WikiPage and throws an exception.

Previously, $title = special page and was used in getParserOutput.
$title is invalid there but Wikibase never used that variable.

Would be good to make sure $title is actually related to the content.

This reverts commit 508a3706d64390a82e232c2f35b0acad616d9f6a.

Change-Id: Ib24515be6e76d3f29e2f9048fbb81e5a25b5857a