Fix Schema:ExternalLinksChange logging if no links are left on page
If there are no links left on the page, we would avoid invoking the spam
blacklist filtering entirely, to avoid having to do blacklist lookups,
etc. However, since we want link removal data, explicitly check for this
scenario and mark all current links as removals, and avoid invoking the
rest of the spam blacklist code.

Bug: T115119
Change-Id: I0bcd5b55594e38c0508b21db2c45e5136123efa0