Make spamblacklist log viewable by logged-in users by default
The main rationale for keeping this log private is to that scanners and things
don't see a huge list of spammy URLs and treat that website as spam itself.

There's no actual private content in here since the regular expressions are
public, so making it available to all users is fine.

Change-Id: I6da473b7cbac2c6fd507bde5c970d701e5ac249e