Fix phan-taint-check false positive
This was causing by echo'ing a value from the database. Since
this script does not use the standard Maintenance class,
phan-taint-check was unable to determine it was a command line
script and thought this was an XSS.

Casting to int caused phan-taint-check to know that the value
was safe, and thus prevents the false positive warning.

Change-Id: Ib786fd591ed50f9b3934efa0c3f06e4c307a81f5