Grant bots the "sboverride" userright by default
It's kind of useless to not give this right to anyone by default since
the whole point of introducing the userright was to use it.

Bots are a good place to start as they're usually pre-vetted and
normally just copying already existing links when being blocked
(e.g. archive bots).

Bug: T313107
Bug: T334344
Change-Id: Ib0ff42ed730d3fe5f5bb6c06f0a87b4959c1f989